update bitcoin spv, add most remaining tests

See merge request !7

+++ contracts/IntegralAuction.sol
@@ -105,11 +105,12 @@ contract IntegralAuction is BringYourOwnWhitelist {
         // Require at least two outputs
         require(_tx.extractNumOutputs() >= 2, 'Must have at least 2 TxOuts');
 
+
         // Submit to SPVStore, get _txid back on success
         bytes memory _header = _headers.slice(0, 80);
         bytes32 _txid = spvStore.validateTransaction(_tx, _proof, _index, _header);
 
-        require(uint(spvStore.getTxOutOutputType(_txid, 1)) == 3, 'TxOut at index 1 must be an OP_RETURN');
+        require(spvStore.getTxOutOutputType(_txid, 1) == 3, 'TxOut at index 1 must be an OP_RETURN');
 
         // Update auction state to CLOSED
         auction.state = AuctionStates.CLOSED;
